+++ src/plugins/shared/index.ts
@@ -1,1 +1,2 @@
 export * from "./DebounceInput";
+export * from "./Spinner";