+++ webpack.config.js
@@ -92,6 +92,7 @@ const config = {
       VERSION: JSON.stringify(version),
       UNSPLASH_API_KEY: JSON.stringify(process.env.UNSPLASH_API_KEY),
       NASA_API_KEY: JSON.stringify(process.env.NASA_API_KEY),
+      TRELLO_API_KEY: JSON.stringify(process.env.TRELLO_API_KEY),
     }),
   ],
   devtool: isWeb || !isProduction ? "source-map" : false,